requirements.txt
elevate>=0.1.3
GitPython>=3.1.14
fonttools>=4.22.0